ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pgvector==0.2.4
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
